requirements.txt
altair
numpy
pandas
pydeck
streamlit
constellate-client
nltk
matplotlib
wordcloud
gensim